requirements.txt
python-telegram-bot==20.5
shazamio==0.4.0.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1